refactor(state): extract derived values (WP5.9.6.6b)
- Create useRegistrationDerived module for useMemo blocks
- Move waitlist CTA computation (firstWaitlistEntry, canFirstGroupPlay, etc.)
- Move memberDatabase initialization
- Update useRegistrationAppState to call useRegistrationDerived
- Add export to barrel file

Main file: 616 → 541 lines (-75 lines, -12%)
Derived module: 140 lines
