+++ src/registration/appHandlers/state/index.js
@@ -2,4 +2,5 @@ export { useRegistrationUiState } from './useRegistrationUiState.js';
 export { useRegistrationRuntime } from './useRegistrationRuntime.js';
 export { useRegistrationDataLayer } from './useRegistrationDataLayer.js';
 export { useRegistrationDomainHooks } from './useRegistrationDomainHooks.js';
+export { useRegistrationDerived } from './useRegistrationDerived.js';
 export { buildRegistrationReturn } from './buildRegistrationReturn.js';
