+++ src/registration/appHandlers/useRegistrationAppState.js
@@ -1,4 +1,4 @@
-import { useEffect, useMemo } from 'react';
+import { useEffect } from 'react';
 
 // Import shared utilities from @lib
 import {
@@ -36,6 +36,9 @@ import { useRegistrationRuntime } from './state/useRegistrationRuntime';
 
 // Data Layer module (WP5.9.6.3)
 import { useRegistrationDataLayer } from './state/useRegistrationDataLayer';
+
+// Derived values module (WP5.9.6.6b)
+import { useRegistrationDerived } from './state/useRegistrationDerived';
 
 // Return object builder (WP5.9.6.6a)
 import { buildRegistrationReturn } from './state/buildRegistrationReturn';
@@ -342,116 +345,13 @@ export function useRegistrationAppState({ isMobileView = false } = {}) {
     onTimeout: applyInactivityTimeoutExitSequence,
   });
 
-  // ===== DERIVED VALUES (useMemo) =====
-
-  // CTA state derived from waitlist and available courts
-  const {
-    firstWaitlistEntry,
-    secondWaitlistEntry,
-    canFirstGroupPlay,
-    canSecondGroupPlay,
-    firstWaitlistEntryData,
-    secondWaitlistEntryData,
-  } = useMemo(() => {
-    const normalizedWaitlist = (data.waitlist || []).map((entry) => ({
-      id: entry.id,
-      position: entry.position,
-      groupType: entry.group?.type,
-      joinedAt: entry.joinedAt,
-      minutesWaiting: entry.minutesWaiting,
-      names: (entry.group?.players || []).map((p) => p.displayName || p.name || 'Unknown'),
-      players: entry.group?.players || [],
-    }));
-
-    const firstGroup = normalizedWaitlist[0] || null;
-    const secondGroup = normalizedWaitlist[1] || null;
-    const gateCount = availableCourts.length;
-
-    const live1 = gateCount >= 1 && firstGroup !== null;
-    const live2 = gateCount >= 2 && secondGroup !== null;
-
-    const first = firstGroup
-      ? { id: firstGroup.id, position: firstGroup.position ?? 1, players: firstGroup.players }
-      : null;
-    const second = secondGroup
-      ? { id: secondGroup.id, position: secondGroup.position ?? 2, players: secondGroup.players }
-      : null;
-
-    return {
-      firstWaitlistEntry: first,
-      secondWaitlistEntry: second,
-      canFirstGroupPlay: !!live1,
-      canSecondGroupPlay: !!live2,
-      firstWaitlistEntryData: first,
-      secondWaitlistEntryData: second,
-    };
-  }, [data.waitlist, availableCourts]);
-
-  // Member database (simplified for autocomplete)
-  const memberDatabase = useMemo(() => {
-    const db = {};
-    const names = [
-      'Novak Djokovic',
-      'Carlos Alcaraz',
-      'Jannik Sinner',
-      'Daniil Medvedev',
-      'Alexander Zverev',
-      'Andrey Rublev',
-      'Casper Ruud',
-      'Hubert Hurkacz',
-      'Taylor Fritz',
-      'Alex de Minaur',
-      'Iga Swiatek',
-      'Aryna Sabalenka',
-      'Coco Gauff',
-      'Elena Rybakina',
-      'Jessica Pegula',
-      'Ons Jabeur',
-      'Marketa Vondrousova',
-      'Karolina Muchova',
-      'Beatriz Haddad Maia',
-      'Petra Kvitova',
-      'Stefanos Tsitsipas',
-      'Felix Auger-Aliassime',
-      'Cameron Norrie',
-      'Karen Khachanov',
-      'Frances Tiafoe',
-      'Tommy Paul',
-      'Lorenzo Musetti',
-      'Ben Shelton',
-      'Nicolas Jarry',
-      'Sebastian Korda',
-      'Madison Keys',
-      'Victoria Azarenka',
-      'Daria Kasatkina',
-      'Belinda Bencic',
-      'Caroline Garcia',
-      'Simona Halep',
-      'Elina Svitolina',
-      'Maria Sakkari',
-      'Liudmila Samsonova',
-      'Zheng Qinwen',
-    ];
-
-    for (let i = 1; i <= CONSTANTS.MEMBER_COUNT; i++) {
-      const id = CONSTANTS.MEMBER_ID_START + i;
-      db[id.toString()] = {
-        familyMembers: [
-          {
-            id: id,
-            name: names[i - 1] || `Player ${i}`,
-            phone: `555-${String(i).padStart(4, '0')}`,
-            ranking: ((i - 1) % 20) + 1,
-            winRate: 0.5 + Math.random() * 0.4,
-          },
-        ],
-        playingHistory: [],
-        lastGame: null,
-      };
-    }
-    return db;
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  // ===== DERIVED VALUES MODULE (WP5.9.6.6b) =====
+  const derived = useRegistrationDerived({
+    data,
+    availableCourts,
+    CONSTANTS,
+    isMobileView,
+  });
 
   // ===== DATA LAYER MODULE (WP5.9.6.3) =====
   // Provides getDataService, loadData and handles board subscription
@@ -549,18 +449,6 @@ export function useRegistrationAppState({ isMobileView = false } = {}) {
     getCourtsOccupiedForClearing,
     guardAddPlayerEarly,
     guardAgainstGroupDuplicate,
-  };
-
-  // ===== GROUP DERIVED VALUES (WP5.9.6.6a2) =====
-  const derived = {
-    isMobileView,
-    canFirstGroupPlay,
-    canSecondGroupPlay,
-    firstWaitlistEntry,
-    secondWaitlistEntry,
-    firstWaitlistEntryData,
-    secondWaitlistEntryData,
-    memberDatabase,
   };
 
   // ===== RETURN ALL STATE AND HELPERS =====
